requirements.txt
streamlit
langchain_core
langchain_community
langchain_ollama
pdfplumber
faiss-cpu
numpy
fpdf